# src/main/resources/openapi.yaml
openapi: 3.1.0
info:
  version: 1.0.0
  title: SpringTech
servers:
  - url: http://localhost:8080

paths:
  /user/create:
    post:
      summary: Create user
      operationId: createUser
      tags:
        - user and orders
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDTO'
      responses:
        '201':
          description: Created was successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /user/{id}:
    get:
      summary: Get user
      operationId: getUsernameById
      tags:
        - user and orders
      parameters:
        - name: id
          in: path
          required: true
          description: ID пользовзателя
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: user was founded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /user/update/{id}:
    put:
      summary: Update user
      operationId: userUpdate
      tags:
        - user and orders
      parameters:
        - name: id
          in: path
          required: true
          description: ID пользовзателя
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDTO'
      responses:
        '200':
          description: User was updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /user/delete/{id}:
    delete:
      summary: Delete User
      operationId: deleteUser
      tags:
        - user and orders
      parameters:
        - name: id
          in: path
          required: true
          description: ID пользовзателя
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: User deleted
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /person/create:
    post:
      summary: Create person
      operationId: createPerson
      tags:
        - person and passport
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PersonDTO'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonDTO'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /person/{id}:
    get:
      summary: Get Person
      operationId: getPerson
      tags:
        - person and passport
      parameters:
        - name: id
          in: path
          required: true
          description: id человека
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: person
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonDTO'
        '404':
          description: Person not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /person/update/{id}:
    put:
      summary: Update person
      operationId: personUpdate
      tags:
        - person and passport
      parameters:
        - name: id
          in: path
          required: true
          description: ID человека
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PersonDTO'
      responses:
        '200':
          description: Person updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonDTO'
        '404':
          description: Person not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /person/delete/{id}:
    delete:
      summary: Delete Person
      operationId: deletePerson
      tags:
        - person and passport
      parameters:
        - name: id
          in: path
          required: true
          description: ID человека
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Person deleted
        '404':
          description: Person not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /actor/create:
    post:
      summary: Create actor
      operationId: createActor
      tags:
        - actors and films
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActorDTO'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActorDTO'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /actor/{id}:
    get:
      summary: Get Actor
      operationId: getActor
      tags:
        - actors and films
      parameters:
        - name: id
          in: path
          required: true
          description: id актера
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: get actor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActorDTO'
        '404':
          description: Actor not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /actor/update/{id}:
    put:
      summary: Update actor
      operationId: actorUpdate
      tags:
        - actors and films
      parameters:
        - name: id
          in: path
          required: true
          description: ID актера
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActorDTO'
      responses:
        '200':
          description: Actor updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActorDTO'
        '404':
          description: Actor not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /actor/delete/{id}:
    delete:
      summary: Delete actor
      operationId: deleteActor
      tags:
        - actors and films
      parameters:
        - name: id
          in: path
          required: true
          description: ID актера
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Actor deleted
        '404':
          description: Actor not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    UserDTO:
      required:
        - username
        - orders
      properties:
        username:
          type: string
          minLength: 2
          maxLength: 30
        orders:
          type: array
          items:
            $ref: '#/components/schemas/OrderDTO'
          default: []
    OrderDTO:
      required:
        - orderDetails
      properties:
        orderDetails:
          type: string
          minLength: 2
          maxLength: 30
          example: Вишня

    PersonDTO:
      required:
        - fullName
        - age
        - passport
      properties:
        fullName:
          type: string
          minLength: 4
          maxLength: 40
          example: Иванов Иван Иванович
        age:
          type: integer
          minimum: 16
          example: 16
        passport:
          $ref: '#/components/schemas/PassportDTO'
    PassportDTO:
      required:
        - passportNumber
      properties:
        passportNumber:
          type: string
          minLength: 6
          maxLength: 6
          example: 123456

    ActorDTO:
      required:
        - name
        - age
      properties:
        name:
          type: string
          minLength: 4
          example: Ivan Ivan Ivanovich
        age:
          type: integer
          minimum: 16
          example: 16
        movies:
          type: array
          items:
            $ref: '#/components/schemas/MovieDTO'
          default:
              name: "Gorko 2"
    MovieDTO:
      required:
        - name
      properties:
        name:
          type: string
          minLength: 4
          example: Gorko 2

    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        code:
          type: integer